+++ .travis.yml
@@ -1,3 +1,3 @@
 language: node_js
 script:
-  - make jshint
+  - make test
