+++ manifest.json
@@ -1,6 +1,6 @@
 {
     "background": {
-        "scripts": ["background.js"],
+        "scripts": ["src/background.js"],
         "persistent": false
     },
     "browser_action": {
